Move P2P unit tests related to peer into a single folder
