+++ elements/lisk-p2p/package.json
@@ -29,8 +29,8 @@
 		"lint": "tslint --format verbose --project .",
 		"lint:fix": "npm run lint -- --fix",
 		"test": "TS_NODE_PROJECT=./test/tsconfig.json nyc mocha test/{,/**/}/*.ts",
-		"test:unit": "TS_NODE_PROJECT=./test/tsconfig.json nyc mocha test/unit/*.ts",
-		"test:integration": "USE_REAL_TIMERS=true TS_NODE_PROJECT=./test/tsconfig.json nyc mocha test/integration/*.ts",
+		"test:unit": "TS_NODE_PROJECT=./test/tsconfig.json nyc mocha test/unit/{,/**/}/*.ts",
+		"test:integration": "USE_REAL_TIMERS=true TS_NODE_PROJECT=./test/tsconfig.json nyc mocha test/integration/{,/**/}/*.ts",
 		"test:watch": "npm test -- --watch",
 		"test:watch:min": "npm run test:watch -- --reporter=min",
 		"test:node": "npm run build:check",
