+++ elements/lisk-p2p/test/unit/peer/base.ts
@@ -15,8 +15,8 @@
 
 import { expect } from 'chai';
 
-import { Peer } from '../../src/peer';
-import { P2PDiscoveredPeerInfo } from '../../src/p2p_types';
+import { Peer } from '../../../src/peer';
+import { P2PDiscoveredPeerInfo } from '../../../src/p2p_types';
 
 describe('peer', () => {
 	const defaultPeerInfo: P2PDiscoveredPeerInfo = {
