+++ elements/lisk-p2p/test/unit/peer/inbound_peer.ts
@@ -15,8 +15,8 @@
 import { expect } from 'chai';
 import { SCServerSocket } from 'socketcluster-server';
 
-import { InboundPeer } from '../../src/peer';
-import { P2PDiscoveredPeerInfo } from '../../src/p2p_types';
+import { InboundPeer } from '../../../src/peer';
+import { P2PDiscoveredPeerInfo } from '../../../src/p2p_types';
 
 describe('inbound_peer', () => {
 	const defaultPeerInfo: P2PDiscoveredPeerInfo = {
